requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
